Add note name to the formatted midi_note parameter
When setting the MIDI note parameter in the parameter menu I found it
a little unintuitive to pick a MIDI note number. This commit adds a
formatter to also display the note name (e.g. `C4`) in brackets after
the note number to hopefully make it easier to pick a note.